requirements.txt
# NOTE: Install these packages separately in this order:
# 1. pip install torch==2.2.2+cu118 --index-url https://download.pytorch.org/whl/cu118
# 2. pip install torchtext==0.17.0
# 3. pip install flash-attn --no-build-isolation

# Core dependencies
numpy>=1.24.3
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.2.2
pandas>=2.0.0
matplotlib>=3.7.1
seaborn>=0.12.2
tqdm>=4.65.0
ipython>=8.0.0


# SCGPT
scgpt>=0.2.4

# For training and evaluation
wandb>=0.15.3
tensorboard>=2.13.0
accelerate>=0.20.3
prefect>=2.0.0

# For data processing
scanpy>=1.9.3
anndata>=0.9.1
scipy>=1.10.1
umap-learn>=0.5.3

# Optional: Alternative attention mechanism if needed
# xformers>=0.0.20
